fix(runtimes): move MergeResourceRequirements to util and drop unused helpers
